worker-SCRAPE/stock-news/x-news/requirements.txt
selenium
webdriver_manager
supabase
pytz
requests
beautifulsoup4
pandas
python-dotenv
feedparser
python-dateutil